+++ lib/shameless/model.rb
@@ -136,6 +136,10 @@ module Shameless
         @base.reload
       end
 
+      def fetch(key, default)
+        @base.fetch(key, default)
+      end
+
       def put_cell(cell_values)
         self.class.put_cell(shardable_value, cell_values)
       end